635 BatchService separation
