+++ coffee-bom/coffee-bom-all/pom.xml
@@ -200,7 +200,7 @@
             <!-- Hibernate core -->
 
             <dependency>
-                <groupId>org.hibernate</groupId>
+                <groupId>org.hibernate.orm</groupId>
                 <artifactId>hibernate-core</artifactId>
                 <version>${version.org.hibernate}</version>
                 <exclusions>
