+++ coffee-bom/coffee-bom-project/pom.xml
@@ -130,6 +130,11 @@
             </dependency>
             <dependency>
                 <groupId>hu.icellmobilsoft.coffee</groupId>
+                <artifactId>coffee-se-util</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>hu.icellmobilsoft.coffee</groupId>
                 <artifactId>coffee-system-jpa</artifactId>
                 <version>${project.version}</version>
             </dependency>
