+++ coffee-jpa/pom.xml
@@ -28,7 +28,7 @@
 
 		<!-- JPA -->
 		<dependency>
-			<groupId>org.hibernate</groupId>
+			<groupId>org.hibernate.orm</groupId>
 			<artifactId>hibernate-core</artifactId>
 		</dependency>
