+++ coffee-model/coffee-model-base/pom.xml
@@ -17,6 +17,10 @@
 			<groupId>hu.icellmobilsoft.coffee</groupId>
 			<artifactId>coffee-se-logging</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>hu.icellmobilsoft.coffee</groupId>
+			<artifactId>coffee-tool</artifactId>
+		</dependency>
 
 		<dependency>
 			<groupId>jakarta.platform</groupId>
@@ -24,7 +28,7 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.hibernate</groupId>
+			<groupId>org.hibernate.orm</groupId>
 			<artifactId>hibernate-core</artifactId>
 		</dependency>
