+++ coffee-se/coffee-se-util/src/test/java/hu/icellmobilsoft/coffee/se/util/string/RandomUtilTest.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package hu.icellmobilsoft.coffee.tool.utils.string;
+package hu.icellmobilsoft.coffee.se.util.string;
 
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Assertions;
