+++ coffee-se/pom.xml
@@ -26,6 +26,7 @@
 		<module>coffee-se-api</module>
 		<module>coffee-se-function</module>
 		<module>coffee-se-logging</module>
+		<module>coffee-se-util</module>
 	</modules>
 
 </project>
