+++ coffee-tool/pom.xml
@@ -41,6 +41,10 @@
 			<groupId>hu.icellmobilsoft.coffee</groupId>
 			<artifactId>coffee-se-function</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>hu.icellmobilsoft.coffee</groupId>
+			<artifactId>coffee-se-util</artifactId>
+		</dependency>
 
 		<dependency>
 			<groupId>hu.icellmobilsoft.coffee</groupId>
